+++ src/logic/TrackMetadata.h
@@ -33,8 +33,8 @@
 struct Cdtext;
 
 // Ours.
-//#include "Frames.h"
-using Frames = qint64;
+#include "Frames.h"
+
 #include <logic/serialization/ISerializable.h>
 #include <future/guideline_helpers.h>
 #include <third_party/libcue/libcue.h>
@@ -68,7 +68,7 @@ public:
 
 	/// Length (? or offset from the beginning of the file?) of the pre-audio gap, in frames.
 	/// @todo This should be offset from start of file, need to check and redocument.  Comes from libcue track_get_zero_pre()
-    Frames m_length_pre_gap {0};
+    Frames m_length_pre_gap_frames {0};
 	/// Start of the audio from the beginning of the file, in frames.
 	/// Exactly the value returned by Libcue track_get_start().
 	/// Corresponds to INDEX 01.
@@ -78,7 +78,7 @@ public:
 	Frames m_length_frames {0};
 	/// Length (? or offset from the beginning of the file?) of the post-audio gap, in frames.
 	/// @todo Determine which it is.  This comes from libcue track_get_zero_post().
-	Frames m_length_post_gap {0};
+	Frames m_length_post_gap_frames {0};
 
 	/// CD-Text "pack type indicators".
 	///